[VPU][Tests] Extract GatherBase to a common part
